refactor: update AbstractLoggerHook to AbstractHookHandler
Changed the way hooks inherits logging functionality: from extension and implementation to only extension of the abstract class which inside of it implements the HookHandlerInterface
